major concurrency changes
I rewrote a lot of what we were using, removed the state_transition lock entirely so the transactions are actually interleaved down to the transfer function. I did have to remove some error checking in the form of the nonce handling. we can likely change this to work concurrently, but we will have to think of what values are acceptable from a concurrent nonce standpoint now that each worker thread has their own "state"